Added a function for multiple substitutions in a template stringi #23

+++ lib/libft/src/ft_newstrs.c
@@ -6,13 +6,13 @@
 /*   By: olegmulko <olegmulko@student.42.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/07/08 21:17:01 by olegmulko         #+#    #+#             */
-/*   Updated: 2020/07/08 22:22:35 by olegmulko        ###   ########.fr       */
+/*   Updated: 2020/07/08 22:57:46 by olegmulko        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-t_string	*ft_newstrs(int m_size)
+t_string	*ft_newstrs(size_t m_size)
 {
 	t_string	*new_string;
 	char		*str;
